rename dir and namespace
