+++ Assets/Rector/Scripts/Cameras/CameraBehaviour.cs
@@ -1,6 +1,6 @@
 using System.Linq;
 using R3;
-using Rector.Nodes;
+using Rector.NodeBehaviours;
 using Unity.Cinemachine;
 using UnityEngine;
