+++ Assets/Rector/Scripts/Cameras/CameraTrackingTargetInputBehaviour.cs
@@ -1,5 +1,5 @@
 using R3;
-using Rector.Nodes;
+using Rector.NodeBehaviours;
 using Unity.Cinemachine;
 using UnityEngine;
