+++ Assets/Rector/Scripts/Editor/VfxAssetReader.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
-using Rector.Nodes;
+using Rector.NodeBehaviours;
 using UnityEditor;
 using UnityEngine;
 using UnityEngine.VFX;
