+++ Assets/Rector/Scripts/NodeComponents/VfxInputBehaviour.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using R3;
-using Rector.Nodes;
+using Rector.NodeBehaviours;
 using UnityEngine;
 using UnityEngine.VFX;
