+++ Assets/Rector/Scripts/RectorSettings.cs
@@ -1,5 +1,5 @@
 ﻿using System;
-using Rector.Nodes;
+using Rector.NodeBehaviours;
 using UnityEngine;
 
 namespace Rector
