+++ Assets/Rector/Scripts/SceneManager.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Threading;
 using Cysharp.Threading.Tasks;
 using R3;
-using Rector.Nodes;
+using Rector.NodeBehaviours;
 using Rector.UI;
 using Rector.UI.Graphs;
 using Rector.UI.Graphs.Nodes;
