+++ Assets/Rector/Scripts/UI/Graphs/Nodes/HudStyleNode.cs
@@ -1,6 +1,6 @@
 using System;
 using R3;
-using Rector.Nodes;
+using Rector.NodeBehaviours;
 using Rector.UI.Graphs.Slots;
 using Rector.UI.Hud;
 using UnityEngine;
